Remove EntitySettings and use custom-data-on-any instead
Add custom data on relationships to define a value and a mode.
In Enforce mode (like original), this overrides user choices
In Default mode it should set the value on the Add/Edit screen but allow user to change. (Not working)

Add API and js for Add/Edit Relationship form to disable other values when enforcing an option.